+++ tests/auth-react/fastapi-server/app.py
@@ -13,28 +13,25 @@
 # under the License.
 import os
 import typing
+from typing import Any, Dict, List, Union
 
 import uvicorn  # type: ignore
 from dotenv import load_dotenv
 from fastapi import Depends, FastAPI
 from fastapi.requests import Request
 from fastapi.responses import JSONResponse, PlainTextResponse
+from httpx import AsyncClient
 from starlette.datastructures import Headers
 from starlette.exceptions import ExceptionMiddleware
 from starlette.middleware.cors import CORSMiddleware
 from starlette.responses import Response
 from starlette.types import ASGIApp
-
-from typing_extensions import Literal
-
-from typing import Any, Dict, List, Union
-
-from httpx import AsyncClient
 from supertokens_python import (InputAppInfo, Supertokens, SupertokensConfig,
                                 get_all_cors_headers, init)
 from supertokens_python.framework.fastapi import Middleware
 from supertokens_python.recipe import (emailpassword, passwordless, session,
-                                       thirdparty, thirdpartyemailpassword)
+                                       thirdparty, thirdpartyemailpassword,
+                                       thirdpartypasswordless)
 from supertokens_python.recipe.emailpassword import EmailPasswordRecipe
 from supertokens_python.recipe.emailpassword.types import InputFormField, User
 from supertokens_python.recipe.emailverification import EmailVerificationRecipe
@@ -52,6 +49,9 @@ from supertokens_python.recipe.thirdparty.types import (
     AccessTokenAPI, AuthorisationRedirectAPI, UserInfo, UserInfoEmail)
 from supertokens_python.recipe.thirdpartyemailpassword import \
     ThirdPartyEmailPasswordRecipe
+from supertokens_python.recipe.thirdpartypasswordless import \
+    ThirdPartyPasswordlessRecipe
+from typing_extensions import Literal
 
 load_dotenv()
 
@@ -167,6 +167,7 @@ class CustomAuth0Provider(Provider):
 def custom_init(contact_method: Union[None, Literal['PHONE', 'EMAIL', 'EMAIL_OR_PHONE']] = None,
                 flow_type: Union[None, Literal['USER_INPUT_CODE', 'MAGIC_LINK', 'USER_INPUT_CODE_AND_MAGIC_LINK']] = None):
     PasswordlessRecipe.reset()
+    ThirdPartyPasswordlessRecipe.reset()
     JWTRecipe.reset()
     EmailVerificationRecipe.reset()
     SessionRecipe.reset()
@@ -174,6 +175,23 @@ def custom_init(contact_method: Union[None, Literal['PHONE', 'EMAIL', 'EMAIL_OR_
     EmailPasswordRecipe.reset()
     ThirdPartyEmailPasswordRecipe.reset()
     Supertokens.reset()
+
+    providers_list: List[Provider] = [
+        Google(
+            client_id=os.environ.get('GOOGLE_CLIENT_ID'),  # type: ignore
+            client_secret=os.environ.get('GOOGLE_CLIENT_SECRET')  # type: ignore
+        ), Facebook(
+            client_id=os.environ.get('FACEBOOK_CLIENT_ID'),  # type: ignore
+            client_secret=os.environ.get('FACEBOOK_CLIENT_SECRET')  # type: ignore
+        ), Github(
+            client_id=os.environ.get('GITHUB_CLIENT_ID'),  # type: ignore
+            client_secret=os.environ.get('GITHUB_CLIENT_SECRET')  # type: ignore
+        ), CustomAuth0Provider(
+            client_id=os.environ.get('AUTH0_CLIENT_ID'),  # type: ignore
+            domain=os.environ.get('AUTH0_DOMAIN'),  # type: ignore
+            client_secret=os.environ.get('AUTH0_CLIENT_SECRET')  # type: ignore
+        )
+    ]
 
     if contact_method is not None and flow_type is not None:
         if contact_method == 'PHONE':
@@ -183,12 +201,26 @@ def custom_init(contact_method: Union[None, Literal['PHONE', 'EMAIL', 'EMAIL_OR_
                 ),
                 flow_type=flow_type
             )
+            thirdpartypasswordless_init = thirdpartypasswordless.init(
+                contact_config=ContactPhoneOnlyConfig(
+                    create_and_send_custom_text_message=save_code_text
+                ),
+                flow_type=flow_type,
+                providers=providers_list
+            )
         elif contact_method == 'EMAIL':
             passwordless_init = passwordless.init(
                 contact_config=ContactEmailOnlyConfig(
                     create_and_send_custom_email=save_code_email
                 ),
                 flow_type=flow_type
+            )
+            thirdpartypasswordless_init = thirdpartypasswordless.init(
+                contact_config=ContactEmailOnlyConfig(
+                    create_and_send_custom_email=save_code_email
+                ),
+                flow_type=flow_type,
+                providers=providers_list
             )
         else:
             passwordless_init = passwordless.init(
@@ -198,12 +230,25 @@ def custom_init(contact_method: Union[None, Literal['PHONE', 'EMAIL', 'EMAIL_OR_
                 ),
                 flow_type=flow_type
             )
+            thirdpartypasswordless_init = thirdpartypasswordless.init(
+                contact_config=ContactEmailOrPhoneConfig(
+                    create_and_send_custom_email=save_code_email,
+                    create_and_send_custom_text_message=save_code_text
+                ),
+                flow_type=flow_type,
+                providers=providers_list
+            )
     else:
         passwordless_init = passwordless.init(
             contact_config=ContactPhoneOnlyConfig(
                 create_and_send_custom_text_message=save_code_text
             ),
             flow_type='USER_INPUT_CODE_AND_MAGIC_LINK'
+        )
+        thirdpartypasswordless_init = thirdpartypasswordless.init(
+            contact_config=ContactPhoneOnlyConfig(create_and_send_custom_text_message=save_code_text),
+            flow_type='USER_INPUT_CODE_AND_MAGIC_LINK',
+            providers=providers_list
         )
 
     recipe_list = [
@@ -218,44 +263,15 @@ def custom_init(contact_method: Union[None, Literal['PHONE', 'EMAIL', 'EMAIL_OR_
             )
         ),
         thirdparty.init(
-            sign_in_and_up_feature=thirdparty.SignInAndUpFeature([
-                Google(
-                    client_id=os.environ.get('GOOGLE_CLIENT_ID'),  # type: ignore
-                    client_secret=os.environ.get('GOOGLE_CLIENT_SECRET')  # type: ignore
-                ), Facebook(
-                    client_id=os.environ.get('FACEBOOK_CLIENT_ID'),  # type: ignore
-                    client_secret=os.environ.get('FACEBOOK_CLIENT_SECRET')  # type: ignore
-                ), Github(
-                    client_id=os.environ.get('GITHUB_CLIENT_ID'),  # type: ignore
-                    client_secret=os.environ.get('GITHUB_CLIENT_SECRET')  # type: ignore
-                ), CustomAuth0Provider(
-                    client_id=os.environ.get('AUTH0_CLIENT_ID'),  # type: ignore
-                    domain=os.environ.get('AUTH0_DOMAIN'),  # type: ignore
-                    client_secret=os.environ.get('AUTH0_CLIENT_SECRET')  # type: ignore
-                )
-            ])
+            sign_in_and_up_feature=thirdparty.SignInAndUpFeature(providers_list)
         ),
         thirdpartyemailpassword.init(
             sign_up_feature=thirdpartyemailpassword.InputSignUpFeature(
                 form_fields),
-            providers=[
-                Google(
-                    client_id=os.environ.get('GOOGLE_CLIENT_ID'),  # type: ignore
-                    client_secret=os.environ.get('GOOGLE_CLIENT_SECRET')  # type: ignore
-                ), Facebook(
-                    client_id=os.environ.get('FACEBOOK_CLIENT_ID'),  # type: ignore
-                    client_secret=os.environ.get('FACEBOOK_CLIENT_SECRET')  # type: ignore
-                ), Github(
-                    client_id=os.environ.get('GITHUB_CLIENT_ID'),  # type: ignore
-                    client_secret=os.environ.get('GITHUB_CLIENT_SECRET')  # type: ignore
-                ), CustomAuth0Provider(
-                    client_id=os.environ.get('AUTH0_CLIENT_ID'),  # type: ignore
-                    domain=os.environ.get('AUTH0_DOMAIN'),  # type: ignore
-                    client_secret=os.environ.get('AUTH0_CLIENT_SECRET')  # type: ignore
-                )
-            ]
+            providers=providers_list
         ),
-        passwordless_init
+        passwordless_init,
+        thirdpartypasswordless_init
     ]
     init(
         supertokens_config=SupertokensConfig('http://localhost:9000'),
@@ -313,7 +329,7 @@ def test_get_device(request: Request):
 
 @app.get("/test/featureFlags")
 def test_feature_flags(request: Request):
-    available = ['passwordless']
+    available = ['passwordless', 'thirdpartypasswordless']
     return JSONResponse({'available': available})